cloud_functions/move/requirements.txt
google-cloud-storage 
